pyproject.toml
[tool.poetry]
name = "discollama"
version = "0.1.0"
description = ""
authors = ["Michael Yang <mxyng@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.1"
redis = "^5.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
